protein_analysis_service/requirements.txt
fastapi>=0.115.0
uvicorn>=0.31.0
biopython>=1.84
httpx>=0.27.2
redis>=5.1.1
pydantic>=2.9.2
pydantic-settings>=2.5.2
python-dotenv>=1.0.1
pytest>=8.3.3
pytest-asyncio>=0.24.0
respx>=0.21.1
